requirements.txt
langchain
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
opencv-python
pytesseract
pillow
